tableHeaderConfig now requried prop and no longer stored on component state in TimePeriodDataTable. getDefautlTableHeadersConfig now hoisted to a service
