+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/TimePeriodDataTable.tsx
@@ -1,3 +1,5 @@
+import React, { memo, forwardRef } from 'react';
+import last from 'lodash/last';
 import cartesian from '@common/lib/utils/cartesian';
 import formatPretty from '@common/lib/utils/number/formatPretty';
 import { TableData } from '@common/services/tableBuilderService';
@@ -8,9 +10,6 @@ import {
   LocationFilter,
 } from '@frontend/modules/table-tool/components/types/filters';
 import TimePeriod from '@frontend/modules/table-tool/components/types/TimePeriod';
-import last from 'lodash/last';
-import sortBy from 'lodash/sortBy';
-import React, { memo, useEffect, useState, forwardRef } from 'react';
 import DataTableCaption from './DataTableCaption';
 import FixedMultiHeaderDataTable from './FixedMultiHeaderDataTable';
 import { TableHeadersFormValues } from './TableHeadersForm';
@@ -24,97 +23,35 @@ interface Props {
   locations: LocationFilter[];
   results: TableData['result'];
   footnotes?: TableData['footnotes'];
-  tableHeadersConfig?: TableHeadersFormValues;
+  tableHeadersConfig: TableHeadersFormValues;
 }
 
 const TimePeriodDataTable = forwardRef<HTMLElement, Props>(
   (props: Props, dataTableRef) => {
-    const {
-      filters,
-      timePeriods,
-      locations,
-      indicators,
-      results,
-      footnotes,
-      tableHeadersConfig,
-    } = props;
-
-    const [tableHeaders, setTableHeaders] = useState<TableHeadersFormValues>({
-      columnGroups:
-        tableHeadersConfig && tableHeadersConfig.columnGroups
-          ? tableHeadersConfig.columnGroups
-          : [],
-      columns:
-        tableHeadersConfig && tableHeadersConfig.columns
-          ? tableHeadersConfig.columns
-          : [],
-      rowGroups:
-        tableHeadersConfig && tableHeadersConfig.rowGroups
-          ? tableHeadersConfig.rowGroups
-          : [],
-      rows:
-        tableHeadersConfig && tableHeadersConfig.rows
-          ? tableHeadersConfig.rows
-          : [],
-    });
-    const removeSiblinglessTotalRows = (
-      categoryFilters: Dictionary<CategoryFilter[]>,
-    ): CategoryFilter[][] => {
-      return Object.values(categoryFilters).filter(filter => {
-        return filter.length > 1 || !filter[0].isTotal;
-      });
-    };
-
-    useEffect(() => {
-      if (
-        tableHeadersConfig &&
-        tableHeadersConfig.columnGroups &&
-        tableHeadersConfig.columns &&
-        tableHeadersConfig.rowGroups &&
-        tableHeadersConfig.rows
-      ) {
-        setTableHeaders(tableHeadersConfig);
-      }
-    }, [tableHeadersConfig]);
-
-    useEffect(() => {
-      const sortedFilters = sortBy(
-        [...removeSiblinglessTotalRows(filters), locations],
-        [options => options.length],
-      );
-
-      const halfwayIndex = Math.floor(sortedFilters.length / 2);
-
-      setTableHeaders({
-        columnGroups: sortedFilters.slice(0, halfwayIndex),
-        rowGroups: sortedFilters.slice(halfwayIndex),
-        columns: timePeriods,
-        rows: indicators,
-      });
-    }, [filters, timePeriods, locations, indicators]);
+    const { results, footnotes, tableHeadersConfig } = props;
 
     const columnHeaders: string[][] = [
-      ...tableHeaders.columnGroups.map(colGroup =>
+      ...tableHeadersConfig.columnGroups.map(colGroup =>
         colGroup.map(group => group.label),
       ),
-      tableHeaders.columns.map(column => column.label),
+      tableHeadersConfig.columns.map(column => column.label),
     ];
 
     const rowHeaders: string[][] = [
-      ...tableHeaders.rowGroups.map(rowGroup =>
+      ...tableHeadersConfig.rowGroups.map(rowGroup =>
         rowGroup.map(group => group.label),
       ),
-      tableHeaders.rows.map(row => row.label),
+      tableHeadersConfig.rows.map(row => row.label),
     ];
 
     const rowHeadersCartesian = cartesian(
-      ...tableHeaders.rowGroups,
-      tableHeaders.rows,
+      ...tableHeadersConfig.rowGroups,
+      tableHeadersConfig.rows,
     );
 
     const columnHeadersCartesian = cartesian(
-      ...tableHeaders.columnGroups,
-      tableHeaders.columns,
+      ...tableHeadersConfig.columnGroups,
+      tableHeadersConfig.columns,
     );
 
     const rows = rowHeadersCartesian.map(rowFilterCombination => {
@@ -184,5 +121,7 @@ const TimePeriodDataTable = forwardRef<HTMLElement, Props>(
     );
   },
 );
+
+TimePeriodDataTable.displayName = 'TimePeriodDataTable';
 
 export default memo(TimePeriodDataTable);
